+++ PyCharm/quadratic_processing.py
